Moved FlowSafeSubject under package node.internal as it is only used by NodeVaultService
