+++ node/src/test/kotlin/net/corda/node/utilities/ObservablesTests.kt
@@ -3,7 +3,7 @@ package net.corda.node.utilities
 import com.google.common.util.concurrent.SettableFuture
 import net.corda.core.internal.bufferUntilSubscribed
 import net.corda.core.internal.tee
-import net.corda.core.utilities.FlowSafeSubject
+import net.corda.node.internal.FlowSafeSubject
 import net.corda.core.internal.FlowSafeSubscriber
 import net.corda.core.internal.OnNextFailedException
 import net.corda.nodeapi.internal.persistence.*
